backend/requirements.txt
fastapi
uvicorn
pydantic
espn-api
pandas
python-dotenv
beautifulsoup4
requests
plotly
scipy
